fix: separated concerns regarding SettingsEvents
