+++ app/src/main/java/org/librefit/ui/screens/settings/SettingsEvent.kt
@@ -1,0 +1,8 @@
+package org.librefit.ui.screens.settings
+
+sealed class SettingsEvent {
+    data object ExportSuccess : SettingsEvent()
+    data object ExportFailed : SettingsEvent()
+    data object ImportSuccess : SettingsEvent()
+    data object ImportFailed : SettingsEvent()
+}
