+++ app/src/main/java/org/librefit/ui/screens/settings/SettingsScreenViewModel.kt
@@ -176,10 +176,3 @@ class SettingsScreenViewModel @Inject constructor(
         }
     }
 }
-
-sealed class SettingsEvent {
-    data object ExportSuccess : SettingsEvent()
-    data object ExportFailed : SettingsEvent()
-    data object ImportSuccess : SettingsEvent()
-    data object ImportFailed : SettingsEvent()
-}
